+++ app/Services/Order/WooCommerceService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services\Order;
 
+use Carbon\Carbon;
 use Exception;
 use GuzzleHttp\Client as GuzzleHttpClient;
 use Illuminate\Contracts\Database\Eloquent\Builder;
@@ -9,71 +10,101 @@ use Illuminate\Contracts\Database\Eloquent\Builder;
 class WooCommerceService
 {
 
-  protected $wooCommerce;
-  protected $guzzleClient;
+  /** @var GuzzleHttpClient The Guzzle HTTP client instance. */
+  private $guzzleClient;
+  private $perPage;
 
+  /**
+   * Initialize the Guzzle HTTP client with WooCommerce API configurations.
+   *
+   * @return void
+   */
   public function __construct()
   {
+    $this->perPage = 100;
     $this->initializeClient();
   }
 
+  /**
+   * Set up the Guzzle HTTP client with WooCommerce API configurations.
+   *
+   * @return void
+   * @throws Exception if the initialization fails.
+   */
   protected function initializeClient()
   {
     try {
       $this->guzzleClient =  new GuzzleHttpClient([
         'base_uri' => config('api.woo_url'),
-        'timeout' => 10,
         'headers' => [
           'Authorization' => 'Basic ' . base64_encode(config('api.woo_key') . ':' . config('api.woo_secret')),
           'Accept' => 'application/json',
         ],
       ]);
-    } catch (Exception $e) {
-      logger()->error('Failed to connect to WooCommerce:' . $e->getMessage());
-      throw $e;
+    } catch (Exception $exception) {
+      logger()->error('Failed to connect to WooCommerce:' . $exception->getMessage());
+      throw $exception;
     }
   }
+
   /**
-   * @param  array  $where
-   * @param  array  $with
-   * @return Builder
+   * Fetch orders from the WooCommerce API.
+   *
+   * @return array|bool Returns an array of orders if successful, or false otherwise.
    */
   public function fetchOrders()
   {
     try {
-      // Check if Guzzle client is initialized
-      if (!$this->guzzleClient) {
-        // Re-initialize client if not initialized
-        $this->initializeClient(); 
-      }
-      // Make request to fetch orders
-      $response = $this->guzzleClient->get('orders?fields=id', ['http_errors' => false]);
-      // Get response status code
-      $statusCode = $response->getStatusCode();
-      // Handle different response status codes
-      switch ($statusCode) {
-        case 200:
-          return json_decode($response->getBody()->getContents(), true);
-        case 401:
-          // Unauthorized, handle accordingly
-          logger()->error('Unauthorized request to fetch orders from WooCommerce');
-          return false;
-        case 404:
-          // Not found, handle accordingly
-          logger()->error('Route Not found on WooCommerce');
-          return false;
-        case 500:
-          logger()->error('Server Error');
-          return false;
-        default:
-          // Unexpected response status code, log and handle accordingly
-          logger()->error('Unexpected response status code: ' . $statusCode);
-          return false;
-      }
-    } catch (\Exception $e) {
+      $page = 1;
+      $allOrders = [];
+
+      // DO loop while the data is not empty for handling pagination
+      do {
+        // Make request to fetch orders
+        $response = $this->fetchData($page);
+        // Get response status code
+        $statusCode = $response->getStatusCode();
+        // Handle different response status codes
+        switch ($statusCode) {
+          case 200:
+            $data =  json_decode($response->getBody()->getContents(), true);
+            $allOrders = array_merge($allOrders, $data);
+            $page++;
+            break;
+          case 401:
+            // Unauthorized, handle accordingly
+            logger()->error('Unauthorized request to fetch orders from WooCommerce');
+            return false;
+          case 404:
+            // Not found, handle accordingly
+            logger()->error('Route Not found on WooCommerce');
+            return false;
+          case 500:
+            logger()->error('Server Error');
+            return false;
+          default:
+            // Unexpected response status code, log and handle accordingly
+            logger()->error('Unexpected response status code: ' . $statusCode);
+            return false;
+        }
+        // while not empty data
+      } while (!empty($data));
+
+      // returning all order data after all pagination
+      return $allOrders;
+    } catch (\Exception $exception) {
       // Log other exceptions
-      logger()->error('Failed to fetch orders from WooCommerce: ' . $e->getMessage());
+      logger()->error('Failed to fetch orders from WooCommerce: ' . $exception->getMessage());
       return false;
     }
+  }
+
+  private function fetchData($page)
+  {
+    // Calculate the date one month ago
+    $oneMonthAgo = Carbon::now()->subMonth()->setTimezone('UTC')->toIso8601ZuluString();
+    // Make request to fetch orders
+    $response = $this->guzzleClient->get('orders?page=' . $page . '&per_page=' . $this->perPage . '&after=' . $oneMonthAgo, ['http_errors' => false]);
+    return $response;
   }
 }